Seperate Container style.
Seperate Container from About Page and HomePage to reuse in Anothor
components.
