+++ src/Container.jsx
@@ -1,0 +1,14 @@
+import styled from '@emotion/styled';
+
+const Container = styled.div({
+  width: '100%',
+  height: '90vh',
+  backgroundColor: 'rgba(0,0,0,0.2)',
+  display: 'flex',
+  flexDirection: 'column',
+  alignItems: 'center',
+  justifyContent: 'center',
+  fontFamily: " 'Playfair Display', serif",
+});
+
+export default Container;
